backend/requirements.txt
anyio==3.6.2
attrs==22.2.0
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
dacite==1.7.0
dependency-injector==4.41.0
dnspython==2.3.0
Eel==0.14.0
electionguard==1.4.0
exceptiongroup==1.1.0
fast-autocomplete==0.9.0
fastapi==0.89.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
gmpy2==2.1.5
google-i18n-address==2.5.2
greenlet==2.0.1
h11==0.14.0
hypothesis==6.65.2
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
JPype1==1.3.0
jsons==1.6.3
jsonschema==4.17.3
Levenshtein==0.20.9
loguru==0.5.3
MarkupSafe==2.1.2
numpy==1.24.1
opencv-python==4.5.4.60
packaging==23.0
pdf2image==1.16.0
Pillow==9.4.0
pluggy==1.0.0
psutil==5.9.4
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-datadir==1.4.1
pytest-mock==3.10.0
pytest-regressions==2.2.0
python-barcode==0.14.0
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==6.0
Random-Word==1.0.11
rapidfuzz==2.13.7
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.1
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
typer==0.4.0
types-python-dateutil==2.8.19.6
typing_extensions==4.4.0
typish==1.9.3
urllib3==1.26.14
uvicorn==0.20.0
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
zxing==0.14
zxingmod==0.19
